+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/trie/diffbased/bonsai/storage/flat/BonsaiFlatDbStrategy.java
@@ -170,4 +170,10 @@ public abstract class BonsaiFlatDbStrategy extends FlatDbStrategy {
         .streamFromKey(ACCOUNT_INFO_STATE, startKeyHash.toArrayUnsafe())
         .map(pair -> new Pair<>(Bytes32.wrap(pair.getKey()), Bytes.wrap(pair.getValue())));
   }
+
+  @Override
+  public FlatDbStrategy contextSafeClone() {
+    // FlatDBStrategies that care about bonsai context changes should override this
+    return this;
+  }
 }