+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/trie/diffbased/common/worldview/DiffBasedWorldState.java
@@ -157,6 +157,7 @@ public abstract class DiffBasedWorldState
 
     boolean success = false;
 
+    this.worldStateKeyValueStorage.getFlatDbStrategy().updateBlockContext(blockHeader);
     final DiffBasedWorldStateKeyValueStorage.Updater stateUpdater =
         worldStateKeyValueStorage.updater();
     Runnable saveTrieLog = () -> {};