+++ ethereum/core/src/test/java/org/hyperledger/besu/ethereum/bonsai/storage/flat/ArchiveFlatDbReaderStrategyTest.java
@@ -12,23 +12,6 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.plugin.services.storage;
+package org.hyperledger.besu.ethereum.bonsai.storage.flat;
 
-/** Supported database storage format */
-public enum DataStorageFormat {
-  /** Original format. Store all tries */
-  FOREST,
-  /** New format. Store one trie, and trie logs to roll forward and backward */
-  BONSAI,
-  /** The option for storing archive data e.g. state at any block */
-  X_BONSAI_ARCHIVE;
-
-  /**
-   * Returns whether the storage format is one of the Bonsai DB formats
-   *
-   * @return true if it is, otherwise false
-   */
-  public boolean isBonsaiFormat() {
-    return this == BONSAI || this == X_BONSAI_ARCHIVE;
-  }
-}
+public class ArchiveFlatDbReaderStrategyTest {}